changes as discussed
