+++ frontend/src/components/ShootDetails/ShootDetailsCard.vue
@@ -33,7 +33,7 @@ SPDX-License-Identifier: Apache-2.0
           <v-list-item-content>
             <v-list-item-subtitle>Cluster Termination</v-list-item-subtitle>
             <v-list-item-title class="d-flex align-center pt-1">
-              <shoot-warnings :shoot-item="shootItem" cluster-expiration-warning small class="mr-1" />
+              <shoot-messages :shoot-item="shootItem" filter="clusterExpirationMessage" small class="mr-1" />
               <span>{{selfTerminationMessage}}</span>
             </v-list-item-title>
           </v-list-item-content>
@@ -51,7 +51,7 @@ SPDX-License-Identifier: Apache-2.0
           </v-list-item-title>
         </v-list-item-content>
         <v-list-item-action class="mx-0" v-if="!isShootMarkedForDeletion">
-          <shoot-warnings :shoot-item="shootItem" k8s-warning />
+          <shoot-messages :shoot-item="shootItem" filter="k8sMessage" />
         </v-list-item-action>
         <v-list-item-action class="mx-0">
           <shoot-version :shoot-item="shootItem"></shoot-version>
@@ -76,7 +76,7 @@ SPDX-License-Identifier: Apache-2.0
           </v-list-item-title>
         </v-list-item-content>
         <v-list-item-action class="mx-0" v-if="!isShootMarkedForDeletion">
-          <shoot-warnings :shoot-item="shootItem" machine-image-warning />
+          <shoot-messages :shoot-item="shootItem" filter="machineImageMessages" />
         </v-list-item-action>
         <v-list-item-action class="mx-0">
           <worker-configuration :shoot-item="shootItem"></worker-configuration>
@@ -202,7 +202,7 @@ import WorkerConfiguration from '@/components/ShootWorkers/WorkerConfiguration'
 import AccessRestrictionsConfiguration from '@/components/ShootAccessRestrictions/AccessRestrictionsConfiguration'
 import PurposeConfiguration from '@/components/PurposeConfiguration'
 import ShootVersion from '@/components/ShootVersion/ShootVersion'
-import ShootWarnings from '@/components/ShootWarnings/ShootWarnings'
+import ShootMessages from '@/components/ShootMessages/ShootMessages'
 import AddonConfiguration from '@/components/ShootAddons/AddonConfiguration'
 import CopyBtn from '@/components/CopyBtn'
 
@@ -226,7 +226,7 @@ export default {
     PurposeConfiguration,
     AddonConfiguration,
     ShootVersion,
-    ShootWarnings,
+    ShootMessages,
     CopyBtn
   },
   props: {
