+++ frontend/src/components/ShootDetails/ShootLifecycleCard.vue
@@ -17,7 +17,7 @@ SPDX-License-Identifier: Apache-2.0
         <v-list-item-content>
           <v-list-item-title>Hibernation</v-list-item-title>
           <v-list-item-subtitle class="d-flex align-center pt-1">
-            <shoot-warnings :shoot-item="shootItem" hibernation-constraint-warning no-hibernation-schedule-warning small class="mr-1" />
+            <shoot-messages :shoot-item="shootItem" filter="noHibernationScheduleWarning hibernationConstraintWarning" small class="mr-1" />
             {{hibernationDescription}}
           </v-list-item-subtitle>
         </v-list-item-content>
@@ -36,7 +36,7 @@ SPDX-License-Identifier: Apache-2.0
         <v-list-item-content>
           <v-list-item-title>Maintenance</v-list-item-title>
           <v-list-item-subtitle class="d-flex align-center pt-1">
-            <shoot-warnings :shoot-item="shootItem" maintenance-constraint-warning small class="mr-1" />
+            <shoot-messages :shoot-item="shootItem" filter="maintenanceConstraintWarning" small class="mr-1" />
             {{maintenanceDescription}}
           </v-list-item-subtitle>
         </v-list-item-content>
@@ -106,7 +106,7 @@ import MaintenanceStart from '@/components/ShootMaintenance/MaintenanceStart'
 import MaintenanceConfiguration from '@/components/ShootMaintenance/MaintenanceConfiguration'
 import ReconcileStart from '@/components/ReconcileStart'
 import RotateKubeconfigStart from '@/components/RotateKubeconfigStart'
-import ShootWarnings from '@/components/ShootWarnings/ShootWarnings'
+import ShootMessages from '@/components/ShootMessages/ShootMessages'
 
 import { isShootHasNoHibernationScheduleWarning } from '@/utils'
 import TimeWithOffset from '@/utils/TimeWithOffset'
@@ -122,7 +122,7 @@ export default {
     DeleteCluster,
     ReconcileStart,
     RotateKubeconfigStart,
-    ShootWarnings
+    ShootMessages
   },
   props: {
     shootItem: {
