+++ frontend/src/components/ShootMessages/K8sExpirationMessage.vue
