+++ frontend/src/components/ShootMessages/NoHibernationScheduleMessage.vue
