+++ frontend/src/mixins/shootItem.js
@@ -132,6 +132,9 @@ export const shootItem = {
     shootDomain () {
       return get(this.shootSpec, 'dns.domain')
     },
+    shootDnsProviders () {
+      return get(this.shootSpec, 'dns.providers')
+    },
     shootHibernationSchedules () {
       return get(this.shootSpec, 'hibernation.schedules', [])
     },
